pyproject.toml
[project]
name = "dpyverification"
version = "0.1.0"
description = ""
authors = [
    {name = "Jurian Beunk", email = "jurian.beunk@deltares.nl"},
    {name = "Aljen Uitbeijerse", email = "aljen.uitbeijerse@vortech.nl"},
    {name = "Cees Voesenek", email = "cees.voesenek@vortech.nl"}
]
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "xarray==2026.1.0",
    "pyyaml==6.0.1",
    "pydantic==2.12.5",
    "netcdf4==1.7.4",
    "importlib-metadata>=7,<9",
    "requests==2.32.3",
    "xskillscore==0.0.26",
    "scores==2.4.0",
    "pydantic-settings==2.12.0",
]

[tool.mypy]
mypy_path = "$MYPY_CONFIG_FILE_DIR/src"
files = ["**/*.py"]
warn_unused_configs = true
python_version = "3.11"
plugins = ["pydantic.mypy", ]
platform = "linux" # even though the development is done on Windows
warn_unused_ignores = true
warn_return_any = true
check_untyped_defs = true
warn_unreachable = true
# Start with strict checking, disable when appropriate for our project
disallow_untyped_defs = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
disallow_any_expr = true
disallow_any_decorated = true
disallow_any_explicit = true
# Options that are not enabled yet, but may turn out to be appropriate for our project
# Move to above this commment when enabled
#explicit_package_bases = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true

[tool.ruff]
target-version = "py310"
line-length = 100
indent-width = 4

[tool.ruff.format]
indent-style = "space"

[tool.ruff.lint]
# Start of with selecting all linting checks, add specific ignores when appropriate for our project
select = ["ALL"]
# Since we are using ruff >=0.5.0, no need to explicitly ignore deprecated rules
ignore = []

[tool.ruff.lint.mccabe]
# Flag errors (`C901`) whenever the complexity level exceeds max-complexity
# Default level is 10, TBD whether this is appropriate for our project
max-complexity = 10

[tool.ruff.lint.pydocstyle]
convention = "numpy"  # what we do with Sauron

[tool.ruff.lint.isort]
known-first-party = ["my-project-name"]  # again, not needed with a 'flat' project layout

[tool.ruff.lint.per-file-ignores]
"**/__init__.py" = [
    "F401",  # imported but unused
]
"tests/**" = [
    "S101",  # use of `assert`
    "SLF001",  # Private member access
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "mypy==1.10.1",
    "pandas-stubs==2.2.2.240603",
    "pre-commit>=4.5.1",
    "pytest==8.2.2",
    "pytest-cov>=7.0.0",
    "pytest-lazy-fixtures==1.3.2",
    "ruff>=0.5.0",
    "types-pyyaml==6.0.12.20240311",
    "types-requests==2.32.0.20240712",
]
